SONAR: create multilateration module and expose it to outside world
