+++ drivers/sub8_sonar/setup.py
@@ -5,7 +5,7 @@ from catkin_pkg.python_setup import generate_distutils_setup
 
 # fetch values from package.xml
 setup_args = generate_distutils_setup(
-    packages=['sub8_sonar'],
+    packages=['sub8_sonar', 'multilateration'],
 )
 
 setup(**setup_args)
