--- drivers/sub8_sonar/sub8_sonar/__init__.py
@@ -1,1 +1,0 @@
-from sonar_driver import Multilaterator
